requirements.txt
crewai
crewai-tools
python-dotenv
datasets